requirements.txt
graia-ariadne
graia-saya
graiax-silkcoder
graiax-silkcoder[ffmpeg]
requests
graia-scheduler
re
random
datetime
configparser
Pillow
pymysql